// src/features/expenses/types/expense.types.ts
export interface Expense {
  idhazine: string;

  name: string;

  startdate: string;
  enddate: string;

  price: string;

  getfrom: string;

  chevahedi: string;

  noemohasebe: string;

  khalibaki: string;

  resid: string;
}

export interface ExpenseSplit {
  idv: string;

  namev: string;

  price: string;

  nafar: string;

  metter: string;

  day: string;

  naghsh: string;
}